move a bunch of code out of sampler.py
